+++ alicloud/resource_alicloud_instance.go
@@ -93,10 +93,9 @@ func resourceAliyunInstance() *schema.Resource {
 			},
 			"io_optimized": &schema.Schema{
 				Type:         schema.TypeString,
-				Optional:     true,
+				Required:     true,
 				ForceNew:     true,
 				ValidateFunc: validateIoOptimized,
-				Computed:     true,
 			},
 
 			"system_disk_category": &schema.Schema{