+++ terraform/examples/alicloud-ecs-image/main.tf
@@ -29,6 +29,8 @@ resource "alicloud_instance" "instance" {
   internet_charge_type = "${var.internet_charge_type}"
   internet_max_bandwidth_out = "${var.internet_max_bandwidth_out}"
 
+  io_optimized = "${var.io_optimized}"
+
   password = "${var.ecs_password}"
 
   allocate_public_ip = "${var.allocate_public_ip}"