+++ terraform/examples/alicloud-ecs-zone-type/main.tf
@@ -26,6 +26,8 @@ resource "alicloud_instance" "instance" {
   internet_charge_type = "${var.internet_charge_type}"
   internet_max_bandwidth_out = "${var.internet_max_bandwidth_out}"
 
+  io_optimized = "${var.io_optimized}"
+
   password = "${var.ecs_password}"
 
   instance_charge_type = "PostPaid"